Make ./semconv into a module
