+++ example/otel-collector/go.mod
@@ -7,6 +7,7 @@ replace (
 	go.opentelemetry.io/otel/exporters/otlp => ../../exporters/otlp
 	go.opentelemetry.io/otel/label => ../../label
 	go.opentelemetry.io/otel/sdk => ../../sdk
+	go.opentelemetry.io/otel/semconv => ../../semconv
 )
 
 require (
@@ -14,5 +15,6 @@ require (
 	go.opentelemetry.io/otel/exporters/otlp v0.16.0
 	go.opentelemetry.io/otel/label v0.16.0
 	go.opentelemetry.io/otel/sdk v0.16.0
+	go.opentelemetry.io/otel/semconv v0.16.0
 	google.golang.org/grpc v1.35.0
 )
