+++ example/prom-collector/go.mod
@@ -8,6 +8,7 @@ replace (
 	go.opentelemetry.io/otel/exporters/otlp => ../../exporters/otlp
 	go.opentelemetry.io/otel/label => ../../label
 	go.opentelemetry.io/otel/sdk => ../../sdk
+	go.opentelemetry.io/otel/semconv => ../../semconv
 )
 
 require (
