+++ sdk/go.mod
@@ -5,6 +5,7 @@ go 1.14
 replace (
 	go.opentelemetry.io/otel => ../
 	go.opentelemetry.io/otel/label => ../label
+	go.opentelemetry.io/otel/semconv => ../semconv
 )
 
 require (
@@ -13,4 +14,5 @@ require (
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/otel v0.16.0
 	go.opentelemetry.io/otel/label v0.16.0
+	go.opentelemetry.io/otel/semconv v0.16.0
 )
